added 4 sample auto w/ 12 seconds left and new tuning values
